Correct string for user profile prefix on Match.com.
On Match.com's site, user profiles are always displayed on Web pages
whose address starts with:

    http://www.match.com/Profile/Display/About/?uid=

According to the official document that describes how Web addresses
work, which is RFC 3986, available at http://tools.ietf.org/html/rfc3986
the path part of a URL (after the domain name) are case sensitive.
Hence, not `/profile/` but rather `/Profile/`. The this part of the PAT
code originates from the FetLife Age/Sex/Location Search userscript,
which we can tell because of its object name of `FL_ASL.getUserProfile`.

Its purpose is to download the HTML source code of a user's profile.
This is later used to extract the user's self-reported physical location
in order to implement the `broadcastNewProximalReports` function (which
doesn't currently work in PAT-ChristianMingle but does in PAT-FetLife).

So, technically speaking, this part of the code is "dead," i.e.,
non-functional. It "should" eventually be fixed. But then again, a lot
of things "should" be different about dating sites, wouldn't you say? ;)