Extract logged in user's Match.com user ID from their login cookie.
Unlike Christian Mingle, Match.com only stores a user's ID inside their
login cookie, not in a global JavaScript variable. In order to obtain
their user ID, we need to extract that information from their login
cookie. This is performed once in the `getCookie()` global utility
function through a series of parsing steps that use regular expressions
to inspect and extract the requested value `n` from the readable cookies
in `document.cookie`, a global variable available to any scripts running
in the current page.

If you're using a supported browser (Chrome or Firefox), you can do this
yourself by right-clicking on a Match.com page and selecting "Inspect
element." Then switch to the JavaScript Console (usually a button
labelled "Console") and typing `document.cookie`, then pressing Enter on
your keyboard. This will show you the value your scripts can read, which
will be a string of `name=value;` pairs. Note that they are terminated
by a semicolon, and note that the regular expression contained within
the `.match()` function inside the `getCookie()` function uses
semicolons to bracket the locally requested value `n` passed into the
funciton. Specifically:

    match('(^|;)?'+n+'=([^;]*)(;|$)')

Since some of Match.com's cookies include subvalues that look like
querystrings, we then call `split()` on the `match()`'ed value in order
to retrieve only the first (`[0]`'th) `split()`'ed result. The ultimate
effect of this function is to provide us an easily-callable way to
retrieve a logged in user's user ID regardless of where we call it:

    getCookie('uid')

This invocation will always return the current user's User ID (or `uid`,
which is what Match.com calls it) whenever we call it in the script.
This is equivalent to PAT-ChristianMingle's call to `uw.my_username`
(since a "username" is what ChristianMingle uses as a unique ID).
Therefore, we replace every call to `uw.my_username` in the entire
script with a call to `getCookie('uid')` to achieve an identical effect.