Correctly extract human readable location data from public profie.
The `FAADE.getLocationFromProfileHtml` function takes a page of HTML
source code as its parameter, parses it into a DOM tree (saved as the
variable called `doc`) and then extracts the user's human readable
location from their profile. This will look something like "Schenectady,
NY, USA" as one example. Unfortunately, Match.com's HTML is very messy,
so we have to do a number of things to actually find the correct string.

First, I loaded up a Match.com profile page. Doesn't matter who. Any
will work. Then I right-clicked on the part of their profile that
reports their location, and selected "Inspect element" in the contextual
menu. This opens the HTML Inspector and revealed that the part of their
profile that reports their location (right under their user handle) is
inside a `<p>` element which is itself the second element child of a
`<div>` element with a `class` attribute value of `basics`. So the first
thing we do is query the DOM for an element whose class is `basics`:

    doc.querySelector('.basics')

From here, we can take a number of approaches; I opted for a simple text
parsing routine. This has the advantage of ignoring any HTML constructs
and focusing only on the visible text inside the `<div class="basics">`
element. Given an HTML node (which is what `querySelector()` returns),
we can access its `textContent` member property to retrieve only the
textual content of that node (and its children):

    doc.querySelector('.basics').textContent

Since Match.com brackets the location information of a profile inside
old-school text pipes (`|`), we can `split()` the entire string returned
by accessing the `.textContent` member of the HTML Node object returned
by `doc.querySelector()` on those pipes; the first (`0`'th) element in
the Array that `.split()` returns will be whatever text exists before
the first pipe, the second (`1`'st) element will be a space character,
then the location string, then a space character, and the third (`2`'nd)
element in the array will be whatever text comes after the second pipe.

    doc.querySelector('.basics').textContent.split('|')

Since we only care about the location string, which we know is the
second element in the array, we immediately access that element:

    doc.querySelector('.basics').textContent.split('|')[1]

To get rid of the surrounding space characters, we just call `.trim()`.

    doc.querySelector('.basics').textContent.split('|')[1].trim()

At this point, we have a comma-separated string like "Schenectady, NY,
USA", but the PAT code ultimately expects a location to be an array of
three elements (City, State, and Country), so we once again `.split()`
the string, this time on commas AND a space:

    doc.querySelector('.basics').textContent.split('|')[1].trim().split(', ')

By splitting on a comma and a space, our `return` value contains clean
strings for city, state, and country values, giving us the complete line
we use to extract and `return` the location information for later use.