Correctly target CreepShield facial recognition result to profile photo.
I was lazy when I created PAT-ChristianMingle as a fork from PAT-FetLife
and I never bothered to change the function names, but the
`FAADE.creepShield.displayOnFetLife` and its sister function,
`FAADE.creepShield.displayError` are both responsible for showing the
results of any CreepShield.com facial recognition match results on the
user's Match.com public profile. The only question for now is, where
should they show up on the user's profile page?

The very first line in both functions determines the answer to that
question. By setting a local variable called `base_el` to the result of
a `document.querySelector()` invocation, we determine where to place the
results of the CreepShield.com search, whether success or error
(respectively). In both cases, it seems to make the most sense to put
those results right underneath the user's primary profile photo (since
that's the image we'll send to CreepShield.com for searching, anyway).
Conveniently, Match.com uniquely identifies that part of the page in
their HTML by using an `<div id="photo-primary">` element. So, we'll
just snag that element. We then access this node's `parentNode` because
of Match.com's particular HTML page structure (i.e., it wraps the
primary photo with another `div` element, which it uses for some
whitespace around the photo, so we can use that, too).