Correctly detect when we have loaded a user's public profile page.
Inside the PAT code's `main()` function, we need to detect when we are
on certain types of pages on Match.com. This line determines whether or
not we've loaded a user's public profile page. If we have, we perform
the functions needed to overlay their user profile with information from
the PAT database.

To check this, the PAT code inspects the browser's URL. This is always
available in the global `window` object's `location` property. That
property, itself, has another property called `href`, which contains
whatever the current URL is, in full; we need the `href` property
specifically because, on Match.com, a user's profile page always
contains a query string that includes their `uid` value (User ID). So,
we access the `window.location.href` property and, as before, perform a
regular expression `.match()` against it to determine where we are.

In this case, we're looking for a public profile, so we look for the
text reading, in English,
"Profile/Display/About/?uid=SOME_USER_ID_VALUE_HERE" and in regular
expression syntax in JavaScript, that looks like this:

    /Profile\/Display\/About\/\?uid=[^&]+/

However, we also want to know WHOSE profile we're viewing, and since the
User ID value is conveniently tucked into the querystring of the page
itself, we can capture that at the same time as we're `.match()`ing,
using JavaScript's regular expression engine's capturing parenthesis.
Thus, the regular expression we use is actually:

    /Profile\/Display\/About\/\?uid=([^&]+)/

Bluntly, this is somewhat redundant code, so we don't actually save our
capture to a variable, here. We'll do that a few lines down, when we
save the `id_in_url` variable.