Correctly extract user handle, user ID, and primary photo from profile.
Once we recognize that we have loaded a user's public profile page,
there are several bits of information that we want from it. Namely, we
want to know the user's handle (which is what Match.com calls what
FetLife called "nickname"), we want to know their internally unique
identifier (their user ID), and we want to know what their profile photo
is, because that's what we'll send to CreepShield.com to ask for a
facial recognition match against the US National Sex Offenders Registry.

These three lines pull out that information, respectively.

Note again that this is as simple as loading up a user's public profile
page in my browser, right-clicking on the parts that I'm interested in,
learning how the HTML identifies those parts, and then translating them
into DOM query selections.

The user's handle is in an `<h2>` element inside the `<div
class="basics">` element, so we just pull out that element's
`.textContent`.

The user's ID is embedded in the page URL, so we extract it from the
address bar.

The user's avatar or profile picture is its own element with a unique
ID (thanks, Match.com!), so we just grab a reference to that node.