Insert "Report" link on user's public profile at correct place.
This part of the code inserts a "Report USER for a consent violation"
link on the given user's public profile page. What's special about PAT
is that it integrates cleanly into whatever website it's deployed
against; in this case, that means we need to use Match.com's own user
interface to construct a report link that looks like it is part of
Match.com itself.

The most sensible place to do this is to insert the link right alongside
the part of the Match.com profile where we can perform other actions,
like starting a chat with this user, or emailing this user, etc.
Match.com's HTML for this part of the page is a `<div>` element with a
class of `user-cta`. Inside that element are more `<div>` elements, each
with a link inside those. On PAT-ChristianMingle, the equivalent HTML
used `<li>` elements inside an element whose `class` value was
`profileComm` (maybe short for "profile communication"?).

In any event, modifying this part of the code from PAT-ChristianMingle
is as simple as using Match.com's HTML conventions over
ChristianMingle.com's. We change the `.profileComm` to `.user-cta` and
we change the `li` to a `div`. That's it.